+++ src/main/java/pl/edu/pg/s165391/musicstore/band/BandService.java
@@ -3,12 +3,15 @@ package pl.edu.pg.s165391.musicstore.band;
 import lombok.NoArgsConstructor;
 import pl.edu.pg.s165391.musicstore.band.model.Band;
 import pl.edu.pg.s165391.musicstore.user.UserService;
+import pl.edu.pg.s165391.musicstore.user.model.User;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import javax.servlet.http.HttpServletRequest;
 import javax.transaction.Transactional;
+import java.security.AccessControlException;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -24,6 +27,9 @@ public class BandService {
     @PersistenceContext
     EntityManager em;
 
+    @Inject
+    private HttpServletRequest securityContext;
+
     public synchronized List<Band> findAllBands() {
         return em.createNamedQuery(Band.Queries.FIND_ALL, Band.class).getResultList();
     }
@@ -34,16 +40,24 @@ public class BandService {
 
     @Transactional
     public synchronized void saveBand(Band band) {
-        if (band.getId() == null) {
-            em.persist(band);
-        } else {
-            em.merge(band);
+        if (securityContext.isUserInRole(User.Roles.ADMIN)) {
+            if (band.getId() == null) {
+                em.persist(band);
+            } else {
+                em.merge(band);
+            }
+            return;
         }
+        throw new AccessControlException("Access denied");
     }
 
     @Transactional
     public void removeBand(Band band) {
-        em.remove(em.merge(band));
+        if (securityContext.isUserInRole(User.Roles.ADMIN)) {
+            em.remove(em.merge(band));
+            return;
+        }
+        throw new AccessControlException("Access denied");
     }
 
 }
